backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
pandas>=1.5.0
numpy>=1.21.0
joblib>=1.2.0
scikit-learn>=1.2.0
lightgbm>=4.0.0
pyyaml>=6.0
optuna>=3.0.0